Fix for Mac OS X compatibility problem.
SF_INFO::frames is of type sf_count_t, which is a 32bit int on most
platforms, but 64bit on Mac OS X. To pass it as S_INTEGER_PARAM_TYPE,
I have to make an explicit cast to int. Note that this is not the
correct thing to do, though. I left a FIXME note to that respect.

+++ src/simage_libsndfile.c
@@ -58,11 +58,19 @@ libsndfile_stream_open(const char * filename, s_stream * stream,
   sf_command (context->file, SFC_SET_NORM_DOUBLE, NULL, SF_TRUE) ;
 
   s_stream_context_set(stream, (void *)context);
+
+  // FIXME: SF_INFO::frames is of type sf_count_t, which is a 32bit
+  // integer on most platforms, but 64bit on others (like Mac OS X), so 
+  // casting context->sfinfo.frames to int is not totally correct - but
+  // necessary, since we do not have a 64bit s_params type. I don't know
+  // why frames cannot be a simple int anyways - Thomas?
+  // 20030108 kyrah
   s_params_set(s_stream_params(stream),
                "samplerate", S_INTEGER_PARAM_TYPE, context->sfinfo.samplerate,
-               "frames",     S_INTEGER_PARAM_TYPE, context->sfinfo.frames,
+               "frames",     S_INTEGER_PARAM_TYPE, (int) context->sfinfo.frames,
                "channels",   S_INTEGER_PARAM_TYPE, context->sfinfo.channels,
                NULL);
+
   return 1;
 }
